Fix clang warnings from Edge compilation

+++ APSICommon/apsi/item.cpp
@@ -80,7 +80,7 @@ Item &Item::operator =(const string &str)
 
 Item &Item::operator =(const Item &assign)
 {
-    for (int i = 0; i < value_.size(); i++)
+    for (size_t i = 0; i < value_.size(); i++)
         value_[i] = assign.value_[i];
     return *this;
 }
@@ -99,13 +99,13 @@ uint64_t item_part(const std::array<apsi::u64, 2>& value_, uint32_t i, uint32_t 
         j1 = (i * split_length) & 0x3F,  // mod 64
         j2 = ((i + 1) * split_length) & 0x3F;  // mod 64
 #ifdef _DEBUG
-    if (split_length > 64 || i2 > value_.size())
+    if (split_length > 64 || i2 > static_cast<int>(value_.size()))
     {
         throw invalid_argument("invalid split_length, or index out of range");
     }
 #endif
     uint64_t mask = (1ULL << split_length) - 1;
-    if ((i1 == i2) || (i2 == value_.size()))
+    if ((i1 == i2) || (i2 == static_cast<int>(value_.size())))
     {
         return (value_[i1] >> j1) & mask;
     }
